Added the classes of the assignments 3 and 4
