+++ src/main/java/bo/usfx/waltervaquera/RowWeights.java
@@ -1,12 +1,13 @@
-package bo.usfx.assignment04;
+package bo.usfx.waltervaquera;
 
 public final class RowWeights {
+
     private RowWeights() {
     }
+
     public static int[] rowWeights(final int[] weights) {
         int team1Weight = 0;
         int team2Weight = 0;
-
         for (int i = 0; i < weights.length; i++) {
             if (i % 2 == 0) {
                 team1Weight += weights[i];
@@ -14,6 +15,6 @@ public final class RowWeights {
                 team2Weight += weights[i];
             }
         }
-        return new int[] {team1Weight, team2Weight};
+        return new int[]{team1Weight, team2Weight};
     }
 }
