--- src/test/java/bo/usfx/assignment03/CollatzTests.java
@@ -1,14 +1,0 @@
-package bo.usfx.assignment03;
-
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-public class CollatzTests {
-    @Test
-    void sampleTests() {
-        Assert.assertEquals(CollatzC.hotpo(1), 0);
-        Assert.assertEquals(CollatzC.hotpo(5), 5);
-        Assert.assertEquals(CollatzC.hotpo(6), 8);
-        Assert.assertEquals(CollatzC.hotpo(23), 15);
-    }
-}
